+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 ## [Unreleased]
 
 ### Maintenance
+- Bump version to 1.0.54 and update API calls to use fetchPostWithRetry
 - version 1.0.46
 
 ### Added